# docs/contracts/openapi-v1.yaml
openapi: "3.0.1"
info:
  title: spring-ai-ascend API v1
  description: "Northbound facade for the spring-ai-ascend agent platform. Stability: stable (W1 surface — adds /v1/runs lifecycle endpoints in addition to W0 /v1/health). All non-health routes require Bearer authentication and the X-Tenant-Id + Idempotency-Key headers (the latter for mutating verbs only)."
  version: "1.0.0-W1"
  contact:
    name: spring-ai-ascend platform team

servers:
  - url: http://localhost:8080
    description: Local dev server

security:
  - bearerAuth: []

paths:
  /v1/health:
    get:
      summary: Platform health check
      description: "Returns UP status, git SHA, DB round-trip latency, and timestamp. Exempt from X-Tenant-Id and Idempotency-Key header requirements (operator probe path); also exempt from bearer auth (permit-list)."
      operationId: getHealth
      tags: [platform]
      security: []
      parameters: []
      responses:
        "200":
          description: Platform is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /v1/runs:
    post:
      summary: Create a run
      description: "Creates a new run for the authenticated tenant. Per Rule 36 / Layer-0 principle P-F (Cursor Flow & Asynchronous Client Boundary) the server returns 202 + TaskCursor envelope immediately; clients poll cursor_url for state. The Idempotency-Key header is mandatory; replays with the same key + body return the original response; same key + different body returns 409 idempotency_body_drift (ADR-0057)."
      operationId: createRun
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRunRequest"
      responses:
        "202":
          description: "Run accepted (PENDING). Cursor Flow per ADR-0070; client polls cursor_url for state."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskCursor"
        "400":
          description: Invalid request (missing tenant context, malformed body)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: "tenant_mismatch (JWT tenant_id claim != X-Tenant-Id header) or jwt_missing_tenant_claim"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: "idempotency_conflict or idempotency_body_drift (Idempotency-Key seen before)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: invalid_run_spec (Bean Validation failure)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /v1/runs/{runId}:
    get:
      summary: Get a run by id
      description: "Returns the current state of a run within the authenticated tenant's scope. Cross-tenant reads return 404 not_found (architect guidance §9.4 'tenant-scope-as-not-found')."
      operationId: getRun
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Run state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResponse"
        "400":
          description: invalid_request (runId not a UUID)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: "not_found (unknown runId OR cross-tenant access)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /v1/runs/{runId}/cancel:
    post:
      summary: Cancel a run
      description: "Transitions a non-terminal run to CANCELLED. Idempotent: cancelling an already-CANCELLED run returns 200 with the current state. Cancelling a SUCCEEDED/FAILED/EXPIRED run returns 409 illegal_state_transition (per RunStateMachine DFA, Rule 20)."
      operationId: cancelRun
      tags: [runs]
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - $ref: "#/components/parameters/IdempotencyKeyHeader"
        - name: runId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Run cancelled (or already CANCELLED — idempotent)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResponse"
        "401":
          description: Missing or invalid Bearer token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: "not_found (unknown runId OR cross-tenant access)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: illegal_state_transition (target run is SUCCEEDED, FAILED, or EXPIRED)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT validated against the configured JWKS endpoint (app.auth.issuer/jwks-uri). MUST carry a tenant_id claim that matches the X-Tenant-Id header (ADR-0056)."

  parameters:
    TenantIdHeader:
      name: X-Tenant-Id
      in: header
      required: true
      description: "Caller-asserted tenant UUID. Cross-checked against the JWT tenant_id claim (ADR-0056 §3)."
      schema:
        type: string
        format: uuid
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: "UUID-shaped idempotency key. Required on mutating verbs (POST/PUT/PATCH). Reused keys with same body return the original response; same key + different body returns 409 idempotency_body_drift (ADR-0057)."
      schema:
        type: string
        format: uuid

  schemas:
    HealthResponse:
      type: object
      required: [status, sha, db_ping_ns, ts]
      properties:
        status:
          type: string
          example: UP
          description: Always "UP" when the endpoint responds.
        sha:
          type: string
          example: "97b0827"
          description: Git commit SHA of the running build, or "dev" if APP_SHA is not set.
        db_ping_ns:
          type: integer
          format: int64
          description: Round-trip latency to Postgres in nanoseconds.
        ts:
          type: string
          format: date-time
          description: Server timestamp of the response (ISO-8601).

    CreateRunRequest:
      type: object
      required: [capabilityName]
      properties:
        capabilityName:
          type: string
          minLength: 1
          maxLength: 128
          description: "Identifier of the capability/agent to invoke. L1 surface; future waves add tool config, budget envelopes, parent run id, etc."

    RunResponse:
      type: object
      required: [runId, status, capabilityName, createdAt, updatedAt]
      properties:
        runId:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, RUNNING, SUSPENDED, SUCCEEDED, FAILED, CANCELLED, EXPIRED]
          description: "Run lifecycle state per RunStateMachine DFA (Rule 20)."
        capabilityName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: "Stable, machine-readable error code (e.g. tenant_mismatch, idempotency_body_drift, illegal_state_transition, not_found, invalid_run_spec)."
            message:
              type: string
              description: Human-readable detail for log/UI consumption.
            details:
              type: object
              additionalProperties: true
              description: Optional structured payload (field errors, retry hints, etc.).

    TaskCursor:
      type: object
      required: [runId, status, cursor_url]
      description: "Cursor Flow envelope returned by long-horizon POST endpoints per Rule 36 / Layer-0 principle P-F (Cursor Flow & Asynchronous Client Boundary). The client polls cursor_url (GET) or subscribes via SSE; the original request never blocks the connection."
      properties:
        runId:
          type: string
          format: uuid
          description: Server-allocated run identifier; stable across the run's lifetime.
        status:
          type: string
          enum: [PENDING, RUNNING, SUSPENDED, SUCCEEDED, FAILED, CANCELLED, EXPIRED]
          description: "Initial run lifecycle state per RunStateMachine DFA (Rule 20). Always PENDING at cursor issuance."
        cursor_url:
          type: string
          format: uri
          description: "Absolute URL the client uses to poll run status (GET) or subscribe to event stream (SSE). Stable for the lifetime of the run."
        webhook_url:
          type: string
          format: uri
          description: "Optional callback URL the platform invokes for blocking checkpoints (Human-in-the-loop intervention points)."

x-stability-tier: stable
x-contract-owner: agent-platform
x-w-note: "W1 surface — adds the run lifecycle endpoints (POST /v1/runs, GET /v1/runs/{runId}, POST /v1/runs/{runId}/cancel). Cancel is a state transition; DELETE is intentionally NOT used for run resources. W2+ additions (streaming Flux<RunEvent>, three-track RunDispatcher) are out of scope at this version."

x-cursor-flow:
  declared_in_w: W1.x
  pattern: "long-horizon POST returns 202 + TaskCursor; client polls cursor_url for state OR subscribes via SSE; webhook callback for blocking checkpoints"
  applies_to: ["/v1/runs"]
  note: "POST /v1/runs activated as cursor flow in W1.x Phase 8 (ADR-0070). Returns 202 + TaskCursor; the controller fires the dispatcher via CompletableFuture.runAsync so the HTTP response is never gated on dispatcher runtime. Integration test: agent-platform RunCursorFlowIT.createReturns202WithCursorWithin200ms (Rule 36.b). SSE subscription + webhook callback are W2 scope."